introduce `linear_args_in_result` + fixes
When building a new MLIR `func.func` + `func.call` for a Julia function that mutates the outer Julia structure of one of its arguments, we still want to restore the original TracedRArray argument to point to a caller-side value instead of callee-side (even if it has dissapeared from traced_arguments). To allow this, `linear_args_in_result` ensures *all* linear_args are returned by the generated function.
This makes the `update_inplace_bad!` test in `buffer_donation.jl` pass.